refactor: temporarily remove json logging and make logging stuff a bit more maintainable
